Create README
(1) Write a program to take 2 numbers from user and calculate sum of all numbers between them.
(2) Write a program to take numbers from user and find the max and min from them (let users choose the No of numbers.
(3) Write a program to take numbers from user and calculate the sum of them (let users choose the No of numbers.
(4) Write a program that take two numbers and multiply them without using * operation.
(5) Write a program that take two numbers and calculate the reminder without using % operation.
(6) Write a program to calculate the power of a number. The number and its power are input from user.
(7) Write a program that reads a positive integer and computes the factorial.
(8) Write a program that reads a positive integer and checks if it is a prime.
(9) Write a program that reads a positive integer and checks if it is a perfect square.
(10) Write a program that reads a positive integer and check if this number is a base of 2 like 1,2,4,8,16,32, 64...
(11) Write a program to sum the digits in a decimal number
145 -> 1+4+5=10.
(12) write a program to take even numbers from user and print the sum of them after each entry if the user enters 2 odd number the program print "bye" and stopped.
(13) You are designing a poster which prints out numbers with a unique style applied to each of them. The styling is based on the number of closed paths or holes present in a giver number. The number of holes that each of the digits from 0 to 9 have are equal to the number of closed paths in the digit. Their values are:
1, 2, 3, 5 and 7 = 0 holes.
0, 4, 6, and 9 = 1 hole.
8 = 2 holes.
Example if number 3824->3 has 0 holes,8 has 2 holes,4 has 1hole sum=0+2+1=3